restructure code to handle various types of datasets
