+++ .gitignore
@@ -4,4 +4,5 @@
 *.iml
 *.xml
 *.pyc
+target
