+++ JavaExtractor/JPredict/src/main/java/JavaExtractor/App.java
@@ -4,14 +4,6 @@ import JavaExtractor.Common.CommandLineValues;
 import JavaExtractor.Common.Common;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.ThreadPoolExecutor;
 import org.kohsuke.args4j.CmdLineException;
 
 public class App {
@@ -37,54 +29,12 @@ public class App {
       ExtractFeaturesTask extractFeaturesTask = new ExtractFeaturesTask(s_CommandLineValues, code);
       extractFeaturesTask.processFile();
     } else if (s_CommandLineValues.Dir != null) {
-      extractDir();
+      switch (s_CommandLineValues.ds) {
+        case CODESEARCHNET:
+          new CodeSearchNetDataset().extractDir(s_CommandLineValues);
+        default:
+          new DefaultDataset().extractDir(s_CommandLineValues);
+      }
     }
-  }
-
-  private static void extractDir() {
-    ThreadPoolExecutor executor =
-        (ThreadPoolExecutor) Executors.newFixedThreadPool(s_CommandLineValues.NumThreads);
-    LinkedList<ExtractFeaturesTask> tasks = new LinkedList<>();
-    try {
-      Files.walk(Paths.get(s_CommandLineValues.Dir))
-          .filter(Files::isRegularFile)
-          .filter(p -> p.toString().toLowerCase().endsWith(".jsonl"))
-          .forEach(
-              f -> {
-                List<String> code;
-                // create file from path (see extractSingleFile on how)
-                try {
-                  code = Files.readAllLines(f);
-                } catch (IOException e) {
-                  e.printStackTrace();
-                  code = new ArrayList<>();
-                }
-                // For each line in file create a new tasks
-                code.forEach(
-                    l -> {
-                      ExtractFeaturesTask task = new ExtractFeaturesTask(s_CommandLineValues, l);
-                      tasks.add(task);
-                    });
-              });
-    } catch (IOException e) {
-      e.printStackTrace();
-      return;
-    }
-    List<Future<Void>> tasksResults = null;
-    try {
-      tasksResults = executor.invokeAll(tasks);
-    } catch (InterruptedException e) {
-      e.printStackTrace();
-    } finally {
-      executor.shutdown();
-    }
-    tasksResults.forEach(
-        f -> {
-          try {
-            f.get();
-          } catch (InterruptedException | ExecutionException e) {
-            // e.printStackTrace();
-          }
-        });
   }
 }
