+++ JavaExtractor/JPredict/src/main/java/JavaExtractor/Common/CommandLineValues.java
@@ -6,11 +6,6 @@ import org.kohsuke.args4j.CmdLineParser;
 import org.kohsuke.args4j.Option;
 import org.kohsuke.args4j.spi.ExplicitBooleanOptionHandler;
 
-enum Dataset {
-  DEFAULT,
-  CODESEARCHNET
-};
-
 /** This class handles the programs arguments. */
 public class CommandLineValues {
   @Option(name = "--file", required = false)
