+++ JavaExtractor/JPredict/src/main/java/JavaExtractor/Common/Dataset.java
@@ -1,0 +1,6 @@
+package JavaExtractor.Common;
+
+public enum Dataset {
+  DEFAULT,
+  CODESEARCHNET
+};
