+++ JavaExtractor/JPredict/src/main/java/JavaExtractor/Dataset.java
@@ -1,0 +1,8 @@
+package JavaExtractor;
+
+import JavaExtractor.Common.CommandLineValues;
+
+interface Dataset {
+
+  public void extractDir(CommandLineValues s_CommandLineValues);
+}
